ss: Bug 475300: Fix inconsistent TreeMapStore due to value comparators
Add Comparable to the ISegment interface for segments to specify
a natural ordering. This ordering will then be used by the tree map
to "break ties" when segments have the same start and end times.

Also removed the getElementAtIndex() method. It was not doing what
the Javadoc said it did, and it does not seem to be useful for the
analysis as it is.

Change-Id: Ic65f50b12f6e94c59d4a60f0e96d9c78ee04b741
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>